Backup.sh
Create a script to backup all files and folders in / home / username:

Zip all files in your home directory.
The archive name must contain the date when the archive was created.
The resulting archive must be saved to the / home / devops folder.
The script should be run using the cron scheduler every Friday at 22:30.