Delete.sh
As we noted, the rm program does not provide room for error and does not provide a mechanism in case you make a mistake or change your mind about deleting a file. Let's fix this. Write a script called delete that will compress the file instead of deleting it (you can choose the gzip or lzma algorithm) and move it to the / home / username / TRASH folder.